Update and enhance settings
