+++ kitchenowl/lib/l10n/app_en.arb
@@ -463,6 +463,7 @@
   "loginTo": "Login to",
   "logout": "Logout",
   "logoutName": "Logout {name}",
+  "longPressToReorder": "Long press to reorder",
   "markAsPaid": "Mark as paid",
   "mealPlanner": "Meal planner",
   "member": "Member",
