+++ kitchenowl/lib/widgets/settings_household/view_settings_list_tile.dart
@@ -24,7 +24,7 @@ class ViewSettingsListTile<Cubit extends HouseholdAddUpdateCubit>
         view.toLocalizedString(context),
       ),
       leading: Icon(view.toIcon(context)),
-      contentPadding: const EdgeInsets.only(left: 20, right: 0),
+      contentPadding: EdgeInsets.only(left: 16),
       trailing: Row(
         mainAxisSize: MainAxisSize.min,
         children: [
