+++ kitchenowl/lib/widgets/shopping_item.dart
@@ -9,6 +9,10 @@ class ShoppingItemWidget<T extends Item> extends StatelessWidget {
   final void Function(T)? onPressed;
   final void Function(T)? onLongPressed;
   final bool selected;
+
+  /// Only applicable if gridStyle = false, raises the list items and makes them fully opaque.
+  /// defaults to true for item is ShoppinglistItem || item is RecipeItem && selected
+  final bool? raised;
   final bool gridStyle;
 
   const ShoppingItemWidget({
@@ -18,6 +22,7 @@ class ShoppingItemWidget<T extends Item> extends StatelessWidget {
     this.onLongPressed,
     this.selected = false,
     this.gridStyle = true,
+    this.raised,
   });
 
   @override
@@ -38,7 +43,8 @@ class ShoppingItemWidget<T extends Item> extends StatelessWidget {
             title: item.name,
             selected: selected,
             icon: ItemIcons.get(item),
-            raised: item is ShoppinglistItem || item is RecipeItem && selected,
+            raised: raised ??
+                item is ShoppinglistItem || item is RecipeItem && selected,
             description: (item is ItemWithDescription)
                 ? (item as ItemWithDescription).description
                 : null,
